refactor(components): resolve circular dependencies | hj
