+++ packages/components/src/utils/dom/isParentOfKind.ts
@@ -1,5 +1,5 @@
 import type { TagName } from '@porsche-design-system/shared';
-import { getTagNameWithoutPrefix } from '..';
+import { getTagNameWithoutPrefix } from '../tag-name';
 
 export const isParentOfKind = (element: HTMLElement, tagName: TagName): boolean => {
   const { parentElement } = element;
