+++ packages/components/src/utils/getNamedSlot.ts
@@ -1,4 +1,5 @@
-import { getHTMLElement, type HTMLElementOrShadowRoot } from './dom';
+import { getHTMLElement } from './dom/getHTMLElement';
+import type { HTMLElementOrShadowRoot } from './dom/dom-types';
 
 export const getNamedSlot = (el: HTMLElementOrShadowRoot, slotName: string): HTMLElement =>
   getHTMLElement(el, `:scope>[slot="${slotName}"]`);
