+++ packages/components/src/utils/jss.ts
@@ -6,7 +6,8 @@ import jssPluginGlobal from 'jss-plugin-global';
 import jssPluginNested from 'jss-plugin-nested';
 import jssPluginSortMediaQueries from 'jss-plugin-sort-css-media-queries';
 import { addImportantToEachRule } from '../styles';
-import { getTagNameWithoutPrefix, hasPropValueChanged } from '.';
+import { hasPropValueChanged } from './has-prop-value-changed';
+import { getTagNameWithoutPrefix } from './tag-name';
 import { type BreakpointCustomizable, parseJSON } from './breakpoint-customizable';
 import { getShadowRootHTMLElement } from './dom';
