+++ packages/components/src/utils/validation/throwIfPropIsUndefined.ts
@@ -1,4 +1,5 @@
-import { getTagNameWithoutPrefix, throwException } from '..';
+import { throwException } from '../log/logger';
+import { getTagNameWithoutPrefix } from '../tag-name';
 
 export const throwIfPropIsUndefined = <T>(element: HTMLElement, propName: string, value: T): void => {
   if (value === undefined) {
